+++ LossTracker/src/LossTracker/Migrations/TrackerContextModelSnapshot.cs
@@ -27,7 +27,7 @@ namespace LossTracker.Migrations
 
                     b.Property<int>("MealId");
 
-                    b.Property<int>("NumberOfServings");
+                    b.Property<double>("NumberOfServings");
 
                     b.Property<int?>("ProfileId");
 
@@ -113,9 +113,9 @@ namespace LossTracker.Migrations
 
                     b.Property<int?>("ProfileId");
 
-                    b.Property<int>("WaistInches");
+                    b.Property<double>("WaistInches");
 
-                    b.Property<int>("WeightLbs");
+                    b.Property<double>("WeightLbs");
 
                     b.HasKey("Id");
                 });
