+++ LossTracker/src/LossTracker/Models/DiaryEntry.cs
@@ -7,7 +7,7 @@ namespace LossTracker.Models
     {
         public int Id { get; set; }
         public DateTime Day { get; set; }
-        public int NumberOfServings { get; set; }
+        public double NumberOfServings { get; set; }
         public int FoodId { get; set; }
         public Food Food { get; set; }
         public int MealId { get; set; }
